Create chapter04.html
This assignment is designed to give you your first hands on exposure to the amazing world of event handing. Now that you've seen some videos about CSS, Javascript and events, it is time to learn by putting it all together into something pretty cool in the browser.

Create an HTML file called chapter04.html which uses CSS, HTML and Javascript to do the following. 

Place 3 BLACK buttons in the browser window which are labeled in white - RED, GREEN, BLUE. Add an event to each button which changes the background color of the window to which ever color label the user clicks on. 

Display a counter on the window for each color and display the number of times that color has been displayed. Label each counter accordingly "RED count:", "BLUE count:", "GREEN count:".

 Whenever the mouse passes over any of the buttons, reverse the colors of the button and the label. So if the label color is black and text is white (initial values) then change that button to white with black text. Or visa versa.  Add a counter next to the other counters which shows the number of times each button has changed colors.